Refactor: Move cell helper functions to qpdk/utils.py to resolve pre-commit structure check failures without adding @gf.cell
