+++ CHANGELOG.md
@@ -5,4 +5,5 @@ All notable changes to this project will be documented in this file.
 ## [0.3.6] - 2026-04-14
 
 ### Fixed
+
 - Initial pre-commit hook fixes.
