+++ qpdk/cells/fluxonium.py
@@ -10,7 +10,6 @@ from gdsfactory.component import Component
 from gdsfactory.typings import ComponentSpec, CrossSectionSpec, LayerSpec
 from klayout.db import DCplxTrans
 
-from qpdk.cells.helpers import add_rect, transform_component
 from qpdk.cells.inductor import meander_inductor
 from qpdk.cells.junction import josephson_junction
 from qpdk.tech import (
@@ -18,6 +17,7 @@ from qpdk.tech import (
     get_etch_section,
     superinductor_cross_section,
 )
+from qpdk.utils import add_rect, transform_component
 
 __all__ = ["fluxonium", "fluxonium_with_bbox"]
